Moved the rows type to a separte types.ts file and fixed the padding on the ApprovedProfilesTable.
